Migrate log window tests
