+++ trview.app.tests/trview.app.tests.vcxproj
@@ -100,7 +100,6 @@
     <ClCompile Include="Windows\LightsWindowManagerTests.cpp" />
     <ClCompile Include="Windows\LightsWindowTests.cpp" />
     <ClCompile Include="Windows\LogWindowManagerTests.cpp" />
-    <ClCompile Include="Windows\LogWindowTests.cpp" />
     <ClCompile Include="Windows\PluginsWindowTests.cpp" />
     <ClCompile Include="Windows\RouteWindowManagerTests.cpp" />
     <ClCompile Include="Windows\ViewerTests.cpp" />
