+++ trview.app.ui.tests/LogWindowTests.h
@@ -1,0 +1,5 @@
+#pragma once
+
+struct ImGuiTestEngine;
+
+void register_log_window_tests(ImGuiTestEngine* engine);
