+++ trview.app.ui.tests/imgui.ini
@@ -57,6 +57,11 @@ Pos=898,29
 Size=588,229
 Collapsed=0
 
+[Window][Log 0]
+Pos=60,60
+Size=520,400
+Collapsed=0
+
 [Table][0x52E825D7,3]
 RefScale=13
 Column 0  Width=42
