+++ trview.app.ui.tests/trview.app.ui.tests.vcxproj
@@ -159,6 +159,7 @@
     <ClInclude Include="GoToTests.h" />
     <ClInclude Include="ItemsWindowTests.h" />
     <ClInclude Include="LevelInfoTests.h" />
+    <ClInclude Include="LogWindowTests.h" />
     <ClInclude Include="pch.h" />
     <ClInclude Include="RoomNavigatorTests.h" />
     <ClInclude Include="SettingsWindowTests.h" />
@@ -216,6 +217,7 @@
     <ClCompile Include="GoToTests.cpp" />
     <ClCompile Include="ItemsWindowTests.cpp" />
     <ClCompile Include="LevelInfoTests.cpp" />
+    <ClCompile Include="LogWindowTests.cpp" />
     <ClCompile Include="pch.cpp">
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">Create</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">Create</PrecompiledHeader>
