+++ trview.app.ui.tests/trview_tests.cpp
@@ -6,6 +6,7 @@
 #include "GoToTests.h"
 #include "ItemsWindowTests.h"
 #include "LevelInfoTests.h"
+#include "LogWindowTests.h"
 #include "RoomNavigatorTests.h"
 #include "SettingsWindowTests.h"
 #include "ViewOptionsTests.h"
@@ -18,6 +19,7 @@ void register_trview_tests(ImGuiTestEngine* engine)
     register_go_to_tests(engine);
     register_items_window_tests(engine);
     register_level_info_tests(engine);
+    register_log_window_tests(engine);
     register_room_navigator_controls_tests(engine);
     register_settings_window_tests(engine);
     register_view_options_tests(engine);
